Reorganise code as a library (nerve)
