+++ Cargo.toml
@@ -18,5 +18,6 @@ name = "github-issues"
 test = true
 doc = false
 
-# [lib]
-# name = nerve
+[lib]
+name = "nerve"
+path = "src/nerve/lib.rs"
